refactor: 비자/마스터카드 검증 함수 및 파일명 수정
